Add case for simple slider hit sounding
"Simple" in this case refers to sliders which have no complex hit sound properties specified (e.g. custom index and edge hit sounds).

In these cases the one hit sound the object was given, is used on all edges, as well as on the slider body (latter part is already accounted for).