Use precise tick time for hit sound detection
In this case I had an approximate tick time of 103975, and a timing line changing the custom index 5 ms later, seemingly not taking effect.

Turns out the game looked at the precise tick time, which was 103974.2, meaning the following line was out of range for hit sound leniency.

This accounts for that edge case.